# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Descripción de este servidor backend",
        "title": "API Backend de la aplicación besigabi",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1/",
    "paths": {
...
        "/api/v1/besigabi/creacionTurnos": {
            "get": {
                "description": "Descripción detallada del endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreacionTurnos"
                ],
                "summary": "Resumen de lo que hace  el endpoint",
                "parameters": [
                    {
                        "description": "Parámetros de la consulta de las ordenes iniciales",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CreacionTurnosRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CreacionTurnosResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un turno en el sistema validando cupos y límites por módulo y origen de forma atómica",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreacionTurnos"
                ],
                "summary": "Crea un turno",
                "parameters": [
                    {
                        "description": "Parámetros para la creación del turno",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CrearTurnoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CrearTurnoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.AccionResponse": {
            "type": "object",
            "properties": {
                "sAccion": {
                    "type": "string"
                }
            }
        },
...
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CompaniaResponse": {
            "description": "Estructura que representa una compañía",
            "type": "object",
            "properties": {
                "codigoCompania": {
                    "type": "string",
                    "example": "C01"
                },
                "idCompania": {
                    "type": "integer",
                    "example": 1
                },
                "nombreCompania": {
                    "type": "string",
                    "example": "Compañía de Prueba"
                }
            }
        },
...
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CreacionTurnosRequest": {
            "description": "Estructura que representa un request de un endpoint de la API",
            "type": "object",
            "properties": {
                "creacionTurnos": {
                    "description": "Incluir descripción del campo",
                    "type": "string",
                    "example": "'Valor_de_ejemplo'"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CreacionTurnosResponse": {
            "description": "Estructura que representa un response de un endpoint de la API",
            "type": "object",
            "properties": {
                "creacionTurnos": {
                    "description": "Incluir descripción del campo",
                    "type": "string",
                    "example": "'Valor_de_ejemplo'"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CrearTurnoRequest": {
            "type": "object",
            "properties": {
                "CodigoModulo": {
                    "type": "string"
                },
                "NombreModulo": {
                    "type": "string"
                },
                "accion": {
                    "type": "string"
                },
                "idCompania": {
                    "type": "integer"
                },
                "idSede": {
                    "type": "integer"
                },
                "idServicio": {
                    "type": "integer"
                },
                "idTipoTurno": {
                    "type": "integer"
                },
                "origen": {
                    "type": "integer"
                },
                "paciente": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.PacienteRequest"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CrearTurnoResponse": {
            "type": "object",
            "properties": {
                "Accion": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.AccionResponse"
                },
                "Compania": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CompaniaResponse"
                },
                "Sede": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.SedeResponse"
                },
                "Servicio": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.TipoServicioResponse"
                },
                "Turno": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.DatosTurnoResponse"
                },
                "paciente": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.PacienteRequest"
                },
                "tipoTurno": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.TipoTurnoResponse"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.DatosTurnoResponse": {
            "type": "object",
            "properties": {
                "fechaTurno": {
                    "type": "string"
                },
                "numeroTurno": {
                    "type": "integer"
                }
            }
        },
...
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.PacienteRequest": {
            "type": "object",
            "properties": {
                "apellido1": {
                    "type": "string"
                },
                "apellido2": {
                    "type": "string"
                },
                "fechaNacimiento": {
                    "type": "string"
                },
                "idPaciente": {
                    "type": "integer"
                },
                "idTipoDocumento": {
                    "type": "integer"
                },
                "nomTipoDocumento": {
                    "type": "string"
                },
                "nombre1": {
                    "type": "string"
                },
                "nombre2": {
                    "type": "string"
                },
                "numeroDocumento": {
                    "type": "string"
                },
                "sexo": {
                    "type": "integer"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.SedeResponse": {
            "description": "Estructura que representa una sede",
            "type": "object",
            "properties": {
                "codSede": {
                    "type": "string",
                    "example": "S01"
                },
                "idSede": {
                    "type": "integer",
                    "example": 1
                },
                "nomSede": {
                    "type": "string",
                    "example": "Sede Central"
                }
            }
        },
...
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.TipoServicioResponse": {
            "description": "Estructura que representa un tipo de servicio",
            "type": "object",
            "properties": {
                "codigoServicio": {
                    "type": "string",
                    "example": "S01"
                },
                "idServicio": {
                    "type": "integer",
                    "example": 1
                },
                "nombreServicio": {
                    "type": "string",
                    "example": "Servicio de Prueba"
                }
            }
        },
...
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.TipoTurnoResponse": {
            "description": "Estructura que representa un tipo de turno",
            "type": "object",
            "properties": {
                "codTipoTurno": {
                    "type": "string",
                    "example": "T01"
                },
                "idTipoTurno": {
                    "type": "integer",
                    "example": 1
                },
                "nomTipoTurno": {
                    "type": "string",
                    "example": "Turno Normal"
                }
            }
        },
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
    }
}